+++ src/test/java/io/confluent/kafka/schemaregistry/utils/TestUtils.java
@@ -93,7 +93,7 @@ public class TestUtils {
     }
   }
 
-  public static long registerSchema(String baseUrl, String schemaString, String subject)
+  public static int registerSchema(String baseUrl, String schemaString, String subject)
       throws IOException {
     RegisterSchemaRequest request = new RegisterSchemaRequest();
     request.setSchema(schemaString);
@@ -102,7 +102,7 @@ public class TestUtils {
         baseUrl, RestUtils.DEFAULT_REQUEST_PROPERTIES, request, subject);
   }
 
-  public static long registerDryRun(String baseUrl, String schemaString, String subject)
+  public static int registerDryRun(String baseUrl, String schemaString, String subject)
       throws IOException {
     RegisterSchemaRequest request = new RegisterSchemaRequest();
     request.setSchema(schemaString);
@@ -135,7 +135,7 @@ public class TestUtils {
    * Register a new schema and verify that it can be found on the expected version.
    */
   public static void registerAndVerifySchema(String baseUrl, String schemaString,
-                                             long expectedId, String subject)
+                                             int expectedId, String subject)
   throws IOException {
     assertEquals("Registering a new schema should succeed",
                  expectedId,
